requirements.txt
streamlit==1.32.0
fastapi==0.110.0
uvicorn==0.29.0
langchain==0.1.16
langchain-community==0.0.38
langgraph==0.0.48
langchain-core==0.1.46
chromadb==0.4.24
sentence-transformers==2.7.0
httpx==0.27.0
requests==2.31.0
python-dotenv==1.0.1
pydantic==2.6.4
faiss-cpu==1.8.0
numpy==1.26.4
pandas==2.2.2
openpyxl==3.1.5
twilio
plotly==5.20.0
python-multipart==0.0.9
aiohttp==3.9.5
tenacity==8.2.3
